Figuring out how to use branches
Same chages as before, just renamed the files to make it slightly
clearer what the changes where
